app/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pytesseract
Pillow
pytest
httpx
python-multipart
prometheus-fastapi-instrumentator